requirements.txt
cffi==1.9.1
click==6.6
cryptography==1.6
enum34==1.1.6
Flask==0.11.1
flask-mongoengine==0.8
Flask-WTF==0.13.1
idna==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mongoengine==0.10.7
Pillow==3.4.2
pyasn1==0.1.9
pycparser==2.17
pymongo==3.3.1
pyOpenSSL==16.2.0
pytesseract==0.1.6
requests==2.11.1
six==1.10.0
Werkzeug==0.11.11
WTForms==2.1
